Module OpamFile.Config

Configuration file: $opam/config

include IO_FILE

type t

File contents

val format_version : OpamVersion.t

val empty : t

Empty file

val write : t typed_file -> t -> unit

Write some contents to a file

val read : t typed_file -> t

Read file contents. Raise an error if the file does not exist.

val read_opt : t typed_file -> t option

Returns None on non-existing file

val safe_read : t typed_file -> t

Read file contents. Return empty if the file does not exist.

val read_from_channel : ?filename:t typed_file -> in_channel -> t

val read_from_string : ?filename:t typed_file -> string -> t

val write_to_channel : ?filename:t typed_file -> out_channel -> t -> unit

val write_to_string : ?filename:t typed_file -> t -> string

val root_version : OpamTypes.opam_version

Current root version

val with_switch : OpamTypes.switch -> t -> t

OCaml switch updates

val with_switch_opt : OpamTypes.switch option -> t -> t

val with_installed_switches : OpamTypes.switch list -> t -> t

The list of switches, ordered by most recently used first

val with_repositories : OpamTypes.repository_name list -> t -> t

Repository updates

val with_opam_version : OpamVersion.t -> t -> t

Update opam-version

val with_opam_root_version : OpamVersion.t -> t -> t

val with_criteria : (OpamTypes.solver_criteria * string) list -> t -> t

val with_best_effort_prefix : string -> t -> t

val with_best_effort_prefix_opt : string option -> t -> t

val with_solver : OpamTypes.arg list -> t -> t

val with_solver_opt : OpamTypes.arg list option -> t -> t

val with_jobs : int -> t -> t

val with_jobs_opt : int option -> t -> t

val with_dl_tool : OpamTypes.arg list -> t -> t

val with_dl_tool_opt : OpamTypes.arg list option -> t -> t

val with_dl_jobs : int -> t -> t

val with_dl_cache : OpamTypes.url list -> t -> t

val with_wrappers : Wrappers.t -> t -> t

val with_global_variables : (OpamTypes.variable * OpamTypes.variable_contents * string) list -> t -> t

val with_eval_variables : (OpamTypes.variable * string list * string) list -> t -> t

val with_validation_hook_opt : OpamTypes.arg list option -> t -> t

val with_default_compiler : OpamTypes.formula -> t -> t

val with_default_invariant : OpamTypes.formula -> t -> t

val with_depext : bool -> t -> t

val with_depext_run_installs : bool -> t -> t

val with_depext_cannot_install : bool -> t -> t

val with_depext_bypass : OpamSysPkg.Set.t -> t -> t

val with_sys_pkg_manager_cmd : OpamTypes.filename OpamStd.String.Map.t -> t -> t

val with_swh_fallback : bool -> t -> t

val with_git_location : OpamTypes.dirname -> t -> t

val with_git_location_opt : OpamTypes.dirname option -> t -> t

val opam_version : t -> OpamTypes.opam_version

Return the opam version

val opam_root_version : t -> OpamTypes.opam_version

Return the opam root version, if not defined returns the default version value 2.1~~previous

val opam_root_version_opt : t -> OpamTypes.opam_version option

Return the opam root version if defined

val repositories : t -> OpamTypes.repository_name list

Return the list of repository

val switch : t -> OpamTypes.switch option

Return the OCaml switch

val previous_switch : t -> OpamTypes.switch option

val installed_switches : t -> OpamTypes.switch list

val jobs : t -> int option

Return the number of jobs defined

val dl_tool : t -> OpamTypes.arg list option

val dl_jobs : t -> int

Return the number of download jobs

val dl_cache : t -> OpamTypes.url list

val criteria : t -> (OpamTypes.solver_criteria * string) list

val best_effort_prefix : t -> string option

val solver : t -> OpamTypes.arg list option

val wrappers : t -> Wrappers.t

val global_variables : t -> (OpamTypes.variable * OpamTypes.variable_contents * string) list

variable, value, docstring

val eval_variables : t -> (OpamTypes.variable * string list * string) list

variable, command, docstring

val validation_hook : t -> OpamTypes.arg list option

val default_compiler : t -> OpamTypes.formula

val default_invariant : t -> OpamTypes.formula

val depext : t -> bool

val depext_run_installs : t -> bool

val depext_cannot_install : t -> bool

val depext_bypass : t -> OpamSysPkg.Set.t

val sys_pkg_manager_cmd : t -> OpamTypes.filename OpamStd.String.Map.t

val swh_fallback : t -> bool

val git_location : t -> OpamTypes.dirname option

val fields : (string * (t, OpamParserTypes.FullPos.value) OpamPp.field_parser) list

val to_list : ?filename:'a typed_file -> t -> (string * OpamParserTypes.FullPos.value) list

All file fields as print-AST, Fields within sections are accessed through dot-separated paths

module BestEffort : BestEffortRead with type t := t

val raw_root_version : 'a typed_file -> OpamVersion.t option

Raw read the config file to extract opam-root-version field value.