Module OpamFile.SwitchExport

An extended version of SwitchSelections that can include full opam files as package "name" {} sections, for storing overlays

type t = {

selections : OpamTypes.switch_selections;

extra_files : string OpamHash.Map.t;

overlays : OPAM.t OpamPackage.Name.Map.t;

}

include IO_FILE with type t := t

val format_version : OpamVersion.t

val empty : t

Empty file

val write : t typed_file -> t -> unit

Write some contents to a file

val read : t typed_file -> t

Read file contents. Raise an error if the file does not exist.

val read_opt : t typed_file -> t option

Returns None on non-existing file

val safe_read : t typed_file -> t

Read file contents. Return empty if the file does not exist.

val read_from_channel : ?filename:t typed_file -> in_channel -> t

val read_from_string : ?filename:t typed_file -> string -> t

val write_to_channel : ?filename:t typed_file -> out_channel -> t -> unit

val write_to_string : ?filename:t typed_file -> t -> string