Module type OpamTypes.GenericPackage

At some point we want to abstract so that the same functions can be used over CUDF and OPAM packages

include OpamParallel.VERTEX

include OpamStd.OrderedType

include Set.OrderedType

type t

The type of the set elements.

val to_string : t -> string

val to_json : t OpamJson.encoder

val of_json : t OpamJson.decoder

include Graph.Sig.COMPARABLE with type t := t

val compare : t -> t -> int

val hash : t -> int

val equal : t -> t -> bool

val name_to_string : t -> string

val version_to_string : t -> string