Module OpamFile.Dot_install

.install files

include IO_FILE

type t

File contents

val format_version : OpamVersion.t

val empty : t

Empty file

val write : t typed_file -> t -> unit

Write some contents to a file

val read : t typed_file -> t

Read file contents. Raise an error if the file does not exist.

val read_opt : t typed_file -> t option

Returns None on non-existing file

val safe_read : t typed_file -> t

Read file contents. Return empty if the file does not exist.

val read_from_channel : ?filename:t typed_file -> in_channel -> t

val read_from_string : ?filename:t typed_file -> string -> t

val write_to_channel : ?filename:t typed_file -> out_channel -> t -> unit

val write_to_string : ?filename:t typed_file -> t -> string

val bin : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

List of files to install in $bin/

val sbin : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

List of files to install in $sbin/

val lib : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

List of files to install in $lib/

val toplevel : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

List of toplevel files

val stublibs : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

C bindings

val share : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

List of architecture-independent files

val share_root : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

List of files under the more general share prefix

val etc : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

List of etc files

val doc : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

List of doc files

val man : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

Man pages

val libexec : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

Executable files under lib/

val lib_root : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

Not relative to the package's lib dir

val libexec_root : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.basename option) list

Not relative to the package's lib dir, and with +x set

val misc : t -> (OpamTypes.basename OpamTypes.optional * OpamTypes.filename) list

List of other files to install