Module OpamFile.Syntax

Helper module for manipulation of the raw syntax (opamfile) format. (the specific file handling modules are derived from this one)

val pp_channel : 'a typed_file -> in_channel -> out_channel -> (unit, OpamParserTypes.FullPos.opamfile) OpamPp.t

val of_channel : 'a typed_file -> in_channel -> OpamParserTypes.FullPos.opamfile

val to_channel : 'a typed_file -> out_channel -> OpamParserTypes.FullPos.opamfile -> unit

val of_string : 'a typed_file -> string -> OpamParserTypes.FullPos.opamfile

val to_string : 'a typed_file -> OpamParserTypes.FullPos.opamfile -> string

val to_string_with_preserved_format : 'a typed_file -> ?format_from:'a typed_file -> ?format_from_string:string -> empty:'a -> ?sections:('a, (string option * OpamParserTypes.FullPos.opamfile_item list) list) OpamFormat.I.fields_def -> fields:('a, OpamParserTypes.FullPos.value) OpamFormat.I.fields_def -> (OpamParserTypes.FullPos.opamfile, OpamTypes.filename * 'a) OpamPp.t -> 'a -> string