Module OpamPath.Switch

Switch related paths

Locations of opam internal dirs and files

val root : t -> OpamTypes.switch -> OpamTypes.dirname

The switch prefix: $opam/$switch

val meta_dirname : string

The name of the subdir of the switch prefix where opam data is stored (".opam-switch")

val meta : t -> OpamTypes.switch -> OpamTypes.dirname

The subdirectory of the prefix where opam data lives: $opam/$switch/.opam-switch

val lock : t -> OpamTypes.switch -> OpamTypes.filename

lock file: $meta/lock

val backup_dir : t -> OpamTypes.switch -> OpamTypes.dirname

The directory where backups are stored for this switch

val backup : t -> OpamTypes.switch -> OpamTypes.switch_selections OpamFile.t

Backup file for state export

val selections : t -> OpamTypes.switch -> OpamTypes.switch_selections OpamFile.t

Switch selections $meta/switch-state

val build : t -> OpamTypes.switch -> OpamTypes.package -> OpamTypes.dirname

Temporary folders used to decompress and compile the corresponding archives: $meta/build/$packages

val remove : t -> OpamTypes.switch -> OpamTypes.package -> OpamTypes.dirname

Temporary folders used to decompress the corresponding archives, used only for package removal $meta/remove/$packages

val build_dir : t -> OpamTypes.switch -> OpamTypes.dirname

Temporary folder: $meta/build

val remove_dir : t -> OpamTypes.switch -> OpamTypes.dirname

Temporary folder: $meta/remove

val install : t -> OpamTypes.switch -> OpamTypes.name -> OpamFile.Dot_install.t OpamFile.t

Installed files for a given package: $meta/install/$name.install

val changes : t -> OpamTypes.switch -> OpamTypes.name -> OpamDirTrack.t OpamFile.t

File registering the changes made by the installation of the given package $meta/install/$name.changes

val install_dir : t -> OpamTypes.switch -> OpamTypes.dirname

Installed files: $meta/install/

val reinstall : t -> OpamTypes.switch -> OpamFile.PkgList.t OpamFile.t

Packages to reinstall on next upgrade: $meta/reinstall

val config_dir : t -> OpamTypes.switch -> OpamTypes.dirname

Configuration folder: $meta/config

val switch_config : t -> OpamTypes.switch -> OpamFile.Switch_config.t OpamFile.t

Global config for the switch: $meta/switch-config

val config : t -> OpamTypes.switch -> OpamTypes.name -> OpamFile.Dot_config.t OpamFile.t

Package-specific configuration file for installed packages: $meta/config/$name.config

val sources_dir : t -> OpamTypes.switch -> OpamTypes.dirname

Clean, uncompressed sources for this switch: $meta/sources/

val sources : t -> OpamTypes.switch -> OpamTypes.package -> OpamTypes.dirname

Clean, uncompressed source directory for this package: $meta/sources/$name.$version/

val extra_files_dir : t -> OpamTypes.switch -> OpamTypes.dirname

Temporary switch-local directory where a by-hash map of extra files may be stored. This is used for switch-imports: $meta/extra-files-cache

val extra_file : t -> OpamTypes.switch -> OpamHash.t -> OpamTypes.filename

Extra file with the given hash from the temporary switch-import cache: $meta/extra-files-cache/HASH

val pinned_package : t -> OpamTypes.switch -> OpamTypes.name -> OpamTypes.dirname

Mirror of the sources for a given pinned package: $meta/sources/$name/ (without version)

val environment : t -> OpamTypes.switch -> OpamFile.Environment.t OpamFile.t

Cached environment updates.

val env_relative_to_prefix : OpamTypes.dirname -> OpamFile.Environment.t OpamFile.t

Like environment, but from the switch prefix dir

val installed_opams : t -> OpamTypes.switch -> OpamTypes.dirname

Directory where the metadata of installed packages is mirrored. $meta/packages/

val installed_opams_cache : t -> OpamTypes.switch -> OpamTypes.filename

Cache file for the mirrored installed package metadata $meta/packages/cache

val installed_package_dir : t -> OpamTypes.switch -> OpamTypes.package -> OpamTypes.dirname

The mirror of the package definition for the given installed package $meta/packages/$name.$version/

val installed_opam : t -> OpamTypes.switch -> OpamTypes.package -> OpamFile.OPAM.t OpamFile.t

The mirror of the opam file for the given installed package $meta/packages/$name.$version/opam

val installed_opam_files_dir : t -> OpamTypes.switch -> OpamTypes.package -> OpamTypes.dirname

Mirror of the extra files attached to the package definitions of installed packages $meta/packages/$name.$version/files/

Locations for the visible part of the installation

module Default : sig ... end

Default config

module DefaultF (L : LAYOUT) : sig ... end

Fuctorised version of Default, for replicating a switch's layout in non-switch contexts

Actual config handling the global-config.config indirections

Package-independent dirs

val get_stdpath : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.std_path -> OpamTypes.dirname

val lib_dir : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.dirname

Library path: $prefix/lib

val stublibs : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.dirname

DLL paths

val toplevel : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.dirname

toplevel path: $prefix/lib/toplevel

val doc_dir : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.dirname

Documentation path: $prefix/doc/

val share_dir : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.dirname

Shared directory: $prefix/share

val etc_dir : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.dirname

Etc directory: $prefix/etc

val man_dir : ?num:string -> t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.dirname

Man pages path: $prefix/man/. The optional num argument will add a manN suffix if specified

val bin : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.dirname

Installed binaries: $prefix/bin

val sbin : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.dirname

Installed system binaries: $prefix/sbin

Package dependent dirs

val lib : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.name -> OpamTypes.dirname

Library path for a given package: $prefix/lib/$name

val doc : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.name -> OpamTypes.dirname

Documentation path for a given package: $prefix/doc/$name

val share : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.name -> OpamTypes.dirname

Share directory for a given package: $prefix/share/$package

val etc : t -> OpamTypes.switch -> OpamFile.Switch_config.t -> OpamTypes.name -> OpamTypes.dirname

Etc directory for a given package: $prefix/etc/$package

module Overlay : sig ... end